+++ src/scripts/clear_processed.sh
@@ -1,0 +1,7 @@
+rm ../../data/processed/eval/*
+rm ../../data/processed/features/*
+rm ../../data/processed/preds/*
+rm ../../data/processed/models/*
+rm ../../data/processed/responses/*
+rm ../../data/processed/*.RDS
+rm ../../data/processed/*.feather
